Week#8/requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
pillow
